# docs/swagger.yaml
basePath: /api/v1
definitions:
  bengkel_internal_http_response.APIError:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/bengkel_internal_http_response.FieldError'
        type: array
      message:
        type: string
    type: object
  bengkel_internal_http_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/bengkel_internal_http_response.APIError'
      meta:
        $ref: '#/definitions/bengkel_internal_http_response.Meta'
    type: object
  bengkel_internal_http_response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      rule:
        type: string
    type: object
  bengkel_internal_http_response.Meta:
    properties:
      last_page:
        type: integer
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
    type: object
...
  internal_http_handler.branchInput:
    properties:
      address:
        type: string
      code:
        maxLength: 30
        type: string
      currency:
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 150
        type: string
      phone:
        maxLength: 30
        type: string
      timezone:
        maxLength: 60
        type: string
    required:
    - code
    - currency
    - is_active
    - name
    - timezone
    type: object
...
  internal_http_handler.roleInput:
    properties:
      code:
        maxLength: 60
        type: string
      description:
        type: string
      name:
        maxLength: 100
        type: string
      permission_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - code
    - name
    - permission_ids
    type: object
...
  internal_http_handler.userInput:
    properties:
      branch_id:
        type: string
      email:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 150
        type: string
      password:
        minLength: 8
        type: string
      phone:
        maxLength: 30
        type: string
      role_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - email
    - is_active
    - name
    - password
    - role_ids
    type: object
  internal_http_handler.userUpdateInput:
    properties:
      branch_id:
        type: string
      email:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 150
        type: string
      password:
        minLength: 8
        type: string
      phone:
        maxLength: 30
        type: string
      role_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - email
    - is_active
    - name
    - role_ids
    type: object
...
info:
  contact: {}
  description: API operasional bengkel, kasir, stok, pembayaran, akuntansi, audit,
    dan CMS.
  title: BengkelOS API
  version: "1.0"
paths:
...
  /branches:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Daftar cabang
      tags:
      - Branches
    post:
      consumes:
      - application/json
      parameters:
      - description: Cabang
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.branchInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Tambah cabang
      tags:
      - Branches
  /branches/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: ID cabang
        in: path
        name: id
        required: true
        type: string
      - description: Cabang
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.branchInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Ubah cabang
      tags:
      - Branches
...
  /permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Katalog permission
      tags:
      - Authorization
...
  /roles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Daftar role beserta permission
      tags:
      - Authorization
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.roleInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Tambah role dan permission
      tags:
      - Authorization
  /roles/{id}:
    delete:
      parameters:
      - description: ID role
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Hapus role yang tidak digunakan
      tags:
      - Authorization
    put:
      consumes:
      - application/json
      parameters:
      - description: ID role
        in: path
        name: id
        required: true
        type: string
      - description: Role
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.roleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Ubah role dan sinkronkan permission
      tags:
      - Authorization
...
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Daftar pengguna dan role
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: Pengguna
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.userInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Tambah pengguna
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: ID pengguna
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Hapus pengguna
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: ID pengguna
        in: path
        name: id
        required: true
        type: string
      - description: Pengguna
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.userUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bengkel_internal_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: Ubah pengguna dan sinkronkan role
      tags:
      - Users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Masukkan "Bearer {access_token}".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"